Added Area Average Background Mathod
A new method for the background substraction from the luminescence peak has been added.

This method consists in selecting an area in the raw hyper spectral image. Then an average is made using the luminescence spectra of all of the pixels inside the selection.
A smoothing of this averaged background noise is then carried in order to erase inconsistent points inside the data.